LIBFCREPO-739. Added completed item logging for update and delete.

+++ plastron/commands/delete.py
@@ -1,5 +1,7 @@
 import logging
-from plastron.util import get_title_string, print_header, print_footer, process_resources
+from email.utils import parsedate_to_datetime
+from plastron.exceptions import RESTAPIException
+from plastron.util import get_title_string, ResourceList, parse_predicate_list
 
 logger = logging.getLogger(__name__)
 
@@ -27,6 +29,11 @@ def configure_cli(subparsers):
         dest='use_transactions'
     )
     parser.add_argument(
+        '--completed',
+        help='file recording the URIs of deleted resources',
+        action='store'
+    )
+    parser.add_argument(
         '-f', '--file',
         help='File containing a list of URIs to delete',
         action='store'
@@ -40,9 +47,6 @@ def configure_cli(subparsers):
 
 class Command:
     def __call__(self, fcrepo, args):
-        if not args.quiet:
-            print_header()
-
         self.repository = fcrepo
         self.repository.test_connection()
         self.dry_run = args.dry_run
@@ -50,22 +54,30 @@ class Command:
         if self.dry_run:
             logger.info('Dry run enabled, no actual deletions will take place')
 
-        process_resources(
-            method=self.delete_item,
+        self.resources = ResourceList(
             repository=self.repository,
             uri_list=args.uris,
             file=args.file,
-            recursive=args.recursive,
+            completed_file=args.completed
+        )
+        self.resources.process(
+            method=self.delete_item,
+            traverse=parse_predicate_list(args.recursive),
             use_transaction=args.use_transactions
         )
 
-        if not args.quiet:
-            print_footer()
-
     def delete_item(self, resource, graph):
+        if resource.uri in self.resources.completed:
+            logger.info(f'Resource {resource.uri} has already been deleted; skipping')
+            return
         title = get_title_string(graph)
         if self.dry_run:
             logger.info(f'Would delete resource {resource} {title}')
         else:
-            self.repository.delete(resource.uri)
-            logger.info(f'Deleted resource {resource} {title}')
+            response = self.repository.delete(resource.uri)
+            if response.status_code == 204:
+                logger.info(f'Deleted resource {resource} {title}')
+                timestamp = parsedate_to_datetime(response.headers['date']).isoformat('T')
+                self.resources.log_completed(resource.uri, title, timestamp)
+            else:
+                raise RESTAPIException(response)
